requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cryptography==46.0.6
ecdsa==0.19.2
fastapi==0.135.3
h11==0.16.0
httptools==0.7.1
idna==3.11
mysql-connector-python==9.6.0
passlib==1.7.4
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.26
PyYAML==6.0.3
requests==2.33.1
rsa==4.9.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
# -------------------------------------------------------------------------
# Phase A + B additions — append to existing requirements.txt
# -------------------------------------------------------------------------
# Alembic drives schema migrations. Pulls SQLAlchemy as a transitive dep
# but SQLAlchemy is NOT used in runtime application code until after
# Phase E — this is intentional, see core/README.md.
alembic==1.13.2
SQLAlchemy==2.0.32
PyMySQL==1.1.1            # Alembic driver — app code still uses mysql.connector

# Static analysis for the tenant-scope test (tests/test_static_org_scope.py)
sqlparse==0.5.1

# Test tooling
pytest==8.3.3
hypothesis==6.112.1
